+++ src/core/instruction/instructions/StrInstructionFX55.java
@@ -12,7 +12,7 @@ public class StrInstructionFX55 extends Instruction {
     @Override
     public void execute() {
         for (int i = 0; i <= lastRegister; i++) {
-            byte value = cpu.getRegisters().getRegister(i);
+            int value = cpu.getRegisters().getRegister(i);
             cpu.getMemory().write((char)(cpu.getI()+i),(char)value);
         }
     }
